water_cloud_body/requirements.txt
chromadb==0.5.11
openai==1.43.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
tenacity==9.0.0
